Docstrings and cleanup

+++ trunk/tovid/libtovid/backend/mpeg2enc.py
@@ -1,16 +1,22 @@
 #! /usr/bin/env python
 # mpeg2enc.py
 
+"""Video and frame encoding using ``mpeg2enc``.
+"""
+
 __all__ = [
     'encode',
     'encode_video',
     'encode_frames',
 ]
 
-import os
+# TODO: Filter out some of the inter-backend dependencies
+
+import os, glob
 from libtovid import log
-from libtovid.cli import Command, Pipe
-from libtovid.backend import mplex, ffmpeg
+from libtovid import utils
+from libtovid import cli
+from libtovid.backend import mplex, mplayer, ffmpeg
 
 def encode(source, target, **kw):
     """Encode a multimedia video using mplayer|yuvfps|mpeg2enc.
@@ -19,6 +25,12 @@ def encode(source, target, **kw):
             Input MediaFile
         target
             Output MediaFile
+        **kw
+            Keyword arguments to customize encoding behavior
+
+    Supported keywords:
+
+        None yet
 
     """
     log.warning("This encoder is very experimental, and may not work.")
@@ -26,11 +38,8 @@ def encode(source, target, **kw):
     outname = target.filename
     # YUV raw video FIFO, for piping video from mplayer to mpeg2enc
     yuvfile = '%s.yuv' % outname
-    try:
+    if os.path.exists(yuvfile):
         os.remove(yuvfile)
-    except:
-        pass
-    os.mkfifo(yuvfile)
     
     # Filenames for intermediate streams (ac3/m2v etc.)
     # Appropriate suffix for audio stream
@@ -44,9 +53,9 @@ def encode(source, target, **kw):
     else:
         videofile = '%s.m2v' % outname
     # Do audio
-    encode_audio(source, audiofile, target)
+    ffmpeg.encode_audio(source, audiofile, target)
     # Do video
-    rip.rip_video(source, yuvfile, target)
+    mplayer.rip_video(source, yuvfile, target)
     encode_video(source, yuvfile, videofile, target)
     # Combine audio and video
     mplex.mux(videofile, audiofile, target)
@@ -70,21 +79,21 @@ def encode_video(source, yuvfile, videofile, target):
     # corresp. to $VID_BITRATE, $MPEG2_QUALITY, $DISC_SIZE, etc.
     # Missing options (compared to tovid)
     # -S 700 -B 247 -b 2080 -v 0 -4 2 -2 1 -q 5 -H -o FILE
-    pipe = Pipe()
+    pipe = cli.Pipe()
 
     # Feed the .yuv file into the pipeline
-    cat = Command('cat', yuvfile)
+    cat = cli.Command('cat', yuvfile)
     pipe.add(cat)
 
     # Adjust framerate if necessary by piping through yuvfps
     if source.fps != target.fps:
         log.info("Adjusting framerate")
-        yuvfps = Command('yuvfps')
-        yuvfps.add('-r', float_to_ratio(target.fps))
+        yuvfps = cli.Command('yuvfps')
+        yuvfps.add('-r', utils.float_to_ratio(target.fps))
         pipe.add(yuvfps)
 
     # Encode the resulting .yuv stream by piping into mpeg2enc
-    mpeg2enc = Command('mpeg2enc')
+    mpeg2enc = cli.Command('mpeg2enc')
     # TV system
     if target.tvsys == 'pal':
         mpeg2enc.add('-F', '3', '-n', 'p')
@@ -115,6 +124,7 @@ def encode_video(source, yuvfile, videofile, target):
 # Frame encoder
 #
 # --------------------------------------------------------------------------
+from libtovid import standard
 
 def encode_frames(imagedir, outfile, format, tvsys, aspect, interlaced=False):
     """Convert an image sequence in the given directory to match a target
@@ -134,14 +144,6 @@ def encode_frames(imagedir, outfile, format, tvsys, aspect, interlaced=False):
     Currently supports JPG and PNG images; input images must already be
     at the desired target resolution.
     """
-    #
-    # Trying to implement interlaced in encoding failed with:
-    #    ++ WARN: [mpeg2enc] Frame height won't split into two equal field pictures...
-    #    ++ WARN: [mpeg2enc] forcing encoding as progressive video
-    # in all cases. Setting -Ip -L1 with pnv2yuv and -I2 in mpeg2enc, etc..
-    # The option is still there, so that compatibility continues, but it
-    # currently does nothing.
-    #
     
     # Use absolute path name
     imagedir = os.path.abspath(imagedir)
@@ -150,34 +152,34 @@ def encode_frames(imagedir, outfile, format, tvsys, aspect, interlaced=False):
     images = glob.glob("%s/*" % imagedir)
     extension = images[0][-3:]
     if extension not in ['jpg', 'png']:
-        raise ValueError, "Image format '%s' isn't currently supported to "\
-              "render video from still frames" % extension
+        raise ValueError("Image format '%s' isn't currently supported to "
+              "render video from still frames" % extension)
     # Make sure remaining image files have the same extension
     for img in images:
         if not img.endswith(extension):
-            raise RuntimeWarning, "%s does not have a .%s extension" %\
-                  (img, extension)
+            raise ValueError("%s does not have a .%s extension" %
+                  (img, extension))
 
-    pipe = Pipe()
+    pipe = cli.Pipe()
 
     # Use jpeg2yuv/png2yuv to stream images
     if extension == 'jpg':
-        jpeg2yuv = Command('jpeg2yuv')
+        jpeg2yuv = cli.Command('jpeg2yuv')
         
         jpeg2yuv.add('-Ip') # Progressive
-            
-        jpeg2yuv.add('-f', '%.3f' % fps(tvsys),
+
+        jpeg2yuv.add('-f', '%.3f' % standard.fps(tvsys),
                      '-j', '%s/%%08d.%s' % (imagedir, extension))
         pipe.add(jpeg2yuv)
     elif extension == 'png':
-        #ls = Command('sh', '-c', 'ls %s/*.png' % imagedir)
-        #xargs = Command('xargs', '-n1', 'pngtopnm')
-        png2yuv = Command('png2yuv')
+        #ls = cli.Command('sh', '-c', 'ls %s/*.png' % imagedir)
+        #xargs = cli.Command('xargs', '-n1', 'pngtopnm')
+        png2yuv = cli.Command('png2yuv')
 
         png2yuv.add('-Ip') # Progressive
             
 
-        png2yuv.add('-f', '%.3f' % fps(tvsys),
+        png2yuv.add('-f', '%.3f' % standard.fps(tvsys),
                     '-j', '%s/%%08d.png' % (imagedir))
         
         pipe.add(png2yuv)
@@ -188,7 +190,7 @@ def encode_frames(imagedir, outfile, format, tvsys, aspect, interlaced=False):
     # TODO: Scale to correct target size using yuvscaler or similar
     
     # Pipe image stream into mpeg2enc to encode
-    mpeg2enc = Command('mpeg2enc')
+    mpeg2enc = cli.Command('mpeg2enc')
 
     # Anyways.
     mpeg2enc.add('-I 0') # Progressive
